Added option to force headers if not found i.o. throwing an exception
If a CSV or Spreadsheet does not contain headers an exception is thrown. The user then has to revert to using column numbers i.o. header names. This change allows the option :force_headers to be set and if set to true will force the library to use the headers as they are given. The headers will be applied to the columns in the order the headers are provided (1st header applied to column 1, 2nd header applied to column 2, etc.).